refactor: Update modal dimensions in list-offers-visitor and offer-list components
The modal dimensions in the list-offers-visitor and offer-list components have been updated to improve the display of the map. The width of the modal has been increased to 1200px in both components, and the height of the modalMap element has been increased to 520px. This change ensures that the map is displayed properly and provides a better user experience.